+++ intelmq/bots/parsers/spamhaus/parser_cert.py
@@ -115,7 +115,7 @@ class SpamhausCERTParserBot(ParserBot):
                 event.add('classification.identifier', 'iot')
                 event.add('event_description.text', 'The infected iot device logged in to a honeypot and issued malicous commands.')
             elif malware == 'iotmirai':
-                event.add('classification.type', 'botnet drone')
+                event.add('classification.type', 'infected system')
                 event.add('classification.identifier', 'mirai')
                 event.add('malware.name', 'mirai')
             elif malware == 'ioturl':
@@ -129,7 +129,7 @@ class SpamhausCERTParserBot(ParserBot):
                 if malware == 'auto':
                     malware = 's_other'
                 event.add('malware.name', malware)
-                event.add('classification.type', 'botnet drone')
+                event.add('classification.type', 'infected system')
                 event.add('source.url', row_splitted[5], raise_failure=False)
 
             # otherwise the same ip, ignore